change the amount from double to decimal

+++ Interface/ITotalCalc.cs
@@ -8,6 +8,6 @@ namespace LendingApp.Interface
 {
     public interface ITotalCalc
     {
-         double CalculateTotal();
+         decimal CalculateTotal();
     }
 }
